requirements.txt
streamlit
langchain
langchain-google-genai
langchain-communities
faiss-cpu
pymupdf
python-dotenv
